Pacify compiler warning
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@102063 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objtools/align_format/align_format_util.hpp
@@ -811,7 +811,7 @@ public:
     ///@param out: ostream to add white space
     ///@param number: the number of white spaces desired
     ///
-    static void AddSpace(CNcbiOstream& out, int number);
+    static void AddSpace(CNcbiOstream& out, size_t number);
 
     ///Return ID for GNL label
     ///@param dtg: dbtag to build label from
